requirements.txt
# Core dependencies
setuptools
astor>=0.8.1
git-python>=1.0.3
inquirer>=3.1.3
pyyaml>=6.0.1
rich>=13.4.2
six>=1.16.0
tokentrim>=0.1.13
wget>=3.2
psutil>=5.9.6
html2image>=2.0.4.3
send2trash>=1.8.2
ipykernel>=6.26.0
jupyter-client>=8.6.0
matplotlib>=3.8.2
toml>=0.10.2
tiktoken>=0.7.0
platformdirs>=4.2.0
pydantic>=2.6.4
google-generativeai>=0.7.1
pyperclip>=1.9.0
yaspin>=3.0.2
python-dotenv>=1.0.0
shortuuid>=1.0.13
litellm>=1.41.26
starlette>=0.37.2
html2text>=2024.2.26
selenium>=4.24.0
webdriver-manager>=4.0.2
anthropic>=0.37.1
pyautogui>=0.9.54
typer>=0.12.5
fastapi>=0.111.0
uvicorn>=0.30.1

# Safe mode (code scanning)
# semgrep>=1.52.0

# Dev
# black>=23.10.1
# isort>=5.12.0
# pre-commit>=3.5.0
# pytest>=7.4.0
